Support for mythtv protocol 26 for live tv
This update introduces the changes necessary to support
livetv with mythtv backend servers that use version 26
of the mythtv protocol. The new protocol does not
use a ring buffer but a chain (collection) of files
that change at programming boundaries. This change
maintains backward compatibility with the older protocol
that used a ring buffer (single file).

This update does not support NFS and will use the mythtv
protocol for any backend that uses version 26 of the
protocol or higher.