Introduction of a packing attribute for the text widget
This update adds a pack attribute to the text widget
for future use in situations where text lines need to
be squished closer together. The current implementation
does not pack very well because it quashes the descents
but this might be avoidable if text is drawn from bottom
to top.