TV guide, multiple fixes and enhancements
Fixed potentially suspicious code that may have been causing
crashes reported by users. Filtered out channels that are not
flagged as visible in the mythconverge database. Fixed boudary
condition with wraparound that did not correctly wrap the
channel list around. Sped up the recorder query so that it
doesn't take 10s of seconds to execute.