Informational message if live tv spawns too slowly
This update adds an informational message informing the user
that they may need to configure DNS or add the backend to
the mvps /etc/hosts file. The message is presented if it
takes longer than 5 seconds to spawn live tv. If live tv does
spawn after the message is presented it will be dismissed and
continue normally.