Improved channel handling in the tvguide
This update fixes two defects related to channel handling in
the tvguide. The first is related to storing channels as
integers where a leading 0 would be lost and the backend
would not understand the channel. The second was a crash
caused by punching in a channel number higher than any
available channel using the numeric keypad.