Text widget text packing changes phase I
This update performs the first phase required to pack text into
a text widget intelligently. It builds a structure of lines
that can subsequnetly be drawn in any order and with any spacing
between them. One bug preventing text from being drawn when it
was exactly the width of the widget was fixed and all strlen
and strcpy functions removed for improved performance.