Auto tune functionality for tvguide
This update adds autotune functionality to the tvguide.
Autotune allows a program in the future to be selected
from the guide and when the time for the program arrives
live tv will automatically tune to that channel. Auto
tune events persist only for the duration of the live
tv session.