Current playing channel fix for the tvguide
This update introduces a workaround to an issue that occurs when
the backend is slow to respond to a channel change. If this occurs
the tvguide will display the old channel as current. Ideally a
notification strategy should be used to update the current channel
this workaround periodically polls the current channel and adjusts
it if it isn't correct.