Fix nasty race conditions which were causing delays on myth protocol when communicating with 0.20 backends.
This fix is only for recording playback at the moment, I haven't had
a chance to look do similar for livetv yet.

The main problem here is that REQUEST_BLOCK will block until we've read
all the data from the data socket. So we have to be careful about
discarding mythtv_file while there's a REQUEST_BLOCK in progress.