Major MClient feature changes:
1) The full screen display now supports both server focus and user focus
when using the title list of what is currently playing.  The server focused
item is displayed in blue and the user focused item is displayed in green.

2) Album covers that are accessible by slimserver are now displayed on the
mvpmc's full screen display.  I store my album covers in ordinary jpeg
format, call them "cover.jpg" and place them in the same directory as the
mp3 from that album.
(Martin helped a lot with this feature)

3) There is a green progress bar at the bottom of the 2 line widget which
displays the progress of the current track as a percentage.

4) There is a green volume bar at the top of the 2 line widget which
display the MClient volume.  This should work for both the mvpmc analog
outputs as well as hardware hacked SPDIF / TOSLINK digital output.