Fix mistakes in audio JIT code. Fix seek/demuxer bug. Reset STC after seek.
Make sure that no old data is fed into the demuxer after a seek (previously
flash frames of old content were making it through post-seek). Always
reset the STC after a seek (if a smallish (approx 5 seconds) seek was
performed the mvp wasn't changing the STC to reflect new video PTS and
instead continuing on with the old values, we now reset the STC to 0 to
force it to pick up the current video PTS). Don't allow video to be paused
for more then 0.5 seconds at a given time.