Improve MClient album cover browser.
The album cover browser state machine has been much improved.  The data
gathered is not subject to errors when skipping between groups of album
covers.  Prior to this fix, it was likely the wrong data or album cover
would be retrieved should the user jump through their collection while a
given screen of album covers was retrieved.

MClient is a music client to SqueezeCenter (a.k.a. SlimServer).