Restrict GUI and OSD to the confines of the user defined viewport
Like some others out there, I recently upgraded to a 16:9 LCD TV, but all my
content is still SD, and mostly 4:3.  I prefer to watch 4:3 content zoomed to
the full width of my TV, while maintaining aspect ratio.  This clips some video
content, but I prefer that to the black bars.  Unfortunately this would also
clip the OSD widgets and some of the menus.

This patch cleans up some widget placements that weren't abiding by the
user-defined viewport setting, with an emphasis on the mythtv screens and the
OSD widgets.  While my motivation was to be able to stay in Zoom:1 mode on my
new TV, I accomplished it by adjusting my viewport, and enforcing that rule.
So this patch is not specific to widescreen TVs or the Zoom:1 mode and should
be generally applicable to all use-case scenarios for mvpmc.

Tested with Zoom:1 setting on 16:9 TV with viewport adjusted to cropped screen.

Signed-off-by: Darren Hart <darren@dvhart.com>

Signed-off-by: Simon Hyde <simon@icedrop.net>