Friendly date and duration in minutes settings and config
I noticed the current date and time formatting wasn't very consistent from
screen to screen, and for some reason I have a really hard time parsing strings
like:

2007-12-07 20:59:00 - 22:01:00

and much prefer strings like:

Fri Dec 07 2007, 10:29 PM (32 Min)
or
Fri Dec 07 2007, 10:29 PM - 11:01 PM

I find when thinking about TV, the day of the week is more of a key for me than
the actual day of the month.  Also, while I care very much when the show
starts, the actual time of the ending isn't as relevant as the duration of the
show (I was constantly subtracting start from end in my head... good practice I
guess, but not really necessary).

Since I know there are still others who prefer the existing format, I've tried
to make my changes configurable, and left the default like it is today.  I have
done a little refactoring, which helps reduce the number of places we do date
formatting (wow - it was kind of all over the place!).  I had to move the
existing/new timestr* functions up earlier in the file since I use it earlier.

I left the mythtv program info screen in the "isoformat" for now.  I'd prefer
to change it, but wanted to get the thoughts of other users/developers first.

I added these features in TV Guide, as that is where the 12 hour setting was,
and it was easy to copy the menu entries and all that.  Tested and confirmed
options work individually as well as together, and the default options match
the current defaults.

The new command line options are long format, and I added them to the usage
info as well.

	--friendly-date
	--duration-minutes

Lastly, I added these two options to the config so they are saved across
restarts and reboots.

Signed-off-by: Darren Hart <darren@dvhart.com>

Signed-off-by: Simon Hyde <simon@icedrop.net>