Author: Roger Heflin (rogerheflin@gmail.com)
Changes to reduce the size of the timestamp structure.

The original structure is 7 - longs, the new structure is 1 short int, and
6 unsigned chars, reducing the structure size by 20 bytes, there are 6
timestamps per recording so this reduces the size of *EACH* recording's
data structure by 120 bytes (average structure size on my testing went from 1000
bytes down to 880 bytes), and assuming we have 3 MB useable (2 copies at
1.5MB each during update), this increases the number of recordngs that can be
used from 1500 to 1700.