Changed refmem counters to use *atomics* were possible, no function 	currently exists to add and subtract arbitrary numbers.
Changed refmem counters to have print protected under a REF_DBG_COUNTERS,
	changed the default DBG level to be REF_DBG_COUNTERS, this can be turned
	off later once things are working better, right now this does not appear to
	work so currently there is a fprintf(stderr, there with the debug print.

Changed timestamps to use %hhu rather than %lu and remote typecast.

Changed indentations in timestamps to be tabs, moved the { up a line to be
	consistent with the style of the rest of the code.

Changed printing of refcounters so all should line up.

Changed the upper/lowercase on the after/befores to be consistent.